refactor: split common utils from test files
